pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "huggingface-hub>=0.30.2",
    "langchain>=0.3.23",
    "langchain-anthropic>=0.3.10",
    "langchain-core>=0.3.52",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.12",
    "numpy>=2.2.4",
    "openai>=1.74.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "transformers>=4.51.3",
]
